+++ models/schema/schema_field.go
@@ -597,7 +597,7 @@ type FileOptions struct {
 	MaxSize   int      `form:"maxSize" json:"maxSize"` // in bytes
 	MimeTypes []string `form:"mimeTypes" json:"mimeTypes"`
 	Thumbs    []string `form:"thumbs" json:"thumbs"`
-	Private   bool     `form:"private" json:"private"`
+	Protected bool     `form:"protected" json:"protected"`
 }
 
 func (o FileOptions) Validate() error {